+++ broker/src/broker/statistics/test.rs
@@ -202,9 +202,9 @@ async fn function_calls() {
 
     // Call 2 functions.
     let reply1 = proxy.call(0, &());
-    let call1 = svc.next_function_call().await.unwrap();
+    let call1 = svc.next_call().await.unwrap();
     let reply2 = proxy.call(0, &());
-    let call2 = svc.next_function_call().await.unwrap();
+    let call2 = svc.next_call().await.unwrap();
     let stats = broker.take_statistics().await.unwrap();
     assert_eq!(stats.messages_sent, 2);
     assert_eq!(stats.messages_received, 2);
@@ -213,7 +213,7 @@ async fn function_calls() {
     assert_eq!(stats.functions_replied, 0);
 
     // Reply 1 function call.
-    call1.reply.ok(&()).unwrap();
+    call1.promise.ok(&()).unwrap();
     reply1.await.unwrap().unwrap();
     let stats = broker.take_statistics().await.unwrap();
     assert_eq!(stats.messages_sent, 1);
@@ -223,7 +223,7 @@ async fn function_calls() {
     assert_eq!(stats.functions_replied, 1);
 
     // Reply 1 function call.
-    call2.reply.ok(&()).unwrap();
+    call2.promise.ok(&()).unwrap();
     reply2.await.unwrap().unwrap();
     let stats = broker.take_statistics().await.unwrap();
     assert_eq!(stats.messages_sent, 1);
@@ -268,9 +268,9 @@ async fn events() {
     assert_eq!(stats.events_sent, 0);
 
     // Emit 3 events on 0.
-    client1.emit_event(svc.id(), 0, &()).unwrap();
-    client1.emit_event(svc.id(), 0, &()).unwrap();
-    client1.emit_event(svc.id(), 0, &()).unwrap();
+    svc.emit_event(0, &()).unwrap();
+    svc.emit_event(0, &()).unwrap();
+    svc.emit_event(0, &()).unwrap();
     client1.sync_broker().await.unwrap();
     let stats = broker.take_statistics().await.unwrap();
     assert_eq!(stats.messages_sent, 7);
@@ -280,9 +280,9 @@ async fn events() {
 
     // Emit 2 events on 0.
     // Emit 1 event on 1.
-    client1.emit_event(svc.id(), 0, &()).unwrap();
-    client1.emit_event(svc.id(), 0, &()).unwrap();
-    client1.emit_event(svc.id(), 1, &()).unwrap();
+    svc.emit_event(0, &()).unwrap();
+    svc.emit_event(0, &()).unwrap();
+    svc.emit_event(1, &()).unwrap();
     client1.sync_broker().await.unwrap();
     let stats = broker.take_statistics().await.unwrap();
     assert_eq!(stats.messages_sent, 5);